requirements.txt
Appium-Flutter-Finder==0.4.0
Appium-Python-Client==2.7.1
async-generator==1.10
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
h11==0.14.0
idna==3.4
outcome==1.2.0
pip==21.3.1
pycparser==2.21
PySocks==1.7.1
selenium==4.7.2
setuptools==60.2.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio-websocket==0.9.2
trio==0.22.0
urllib3==1.26.13
wheel==0.37.1
wsproto==1.2.0
